Replace ExperimentItemCollection.java with a vaadin listbox
