+++ vaadinfrontend/frontend/themes/datamanager/components/div.css
@@ -40,19 +40,6 @@
   flex-flow: column;
 }
 
-.experiment-item-collection .content {
-  display: grid;
-  gap: var(--lumo-space-m);
-  grid-auto-rows: max-content;
-}
-
-.experiment-item-collection .header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: var(--lumo-space-m);
-}
-
 .project-collection {
   display: flex;
   gap: 1rem;
