+++ vaadinfrontend/frontend/themes/datamanager/components/page-area.css
@@ -120,6 +120,19 @@
   max-height: 50%;
 }
 
+.page-area.experiment-list-component .content {
+  display: grid;
+  gap: var(--lumo-space-m);
+  grid-auto-rows: max-content;
+}
+
+.page-area.experiment-list-component .header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: var(--lumo-space-m);
+}
+
 .page-area.project-access-component {
   width: 100%;
   height: 100%;
