+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/projects/project/experiments/ExperimentListComponent.java
@@ -1,6 +1,13 @@
 package life.qbic.datamanager.views.projects.project.experiments;
 
+import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.ComponentEvent;
 import com.vaadin.flow.component.ComponentEventListener;
+import com.vaadin.flow.component.button.Button;
+import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.html.Span;
+import com.vaadin.flow.component.listbox.ListBox;
+import com.vaadin.flow.data.renderer.ComponentRenderer;
 import com.vaadin.flow.spring.annotation.SpringComponent;
 import com.vaadin.flow.spring.annotation.UIScope;
 import java.io.Serial;
@@ -8,44 +15,42 @@ import java.util.Collection;
 import java.util.Objects;
 import life.qbic.application.commons.ApplicationException;
 import life.qbic.datamanager.views.Context;
+import life.qbic.datamanager.views.general.Disclaimer;
 import life.qbic.datamanager.views.general.PageArea;
 import life.qbic.datamanager.views.projects.project.experiments.experiment.ExperimentInformationDialog;
-import life.qbic.datamanager.views.support.experiment.ExperimentItem;
-import life.qbic.datamanager.views.support.experiment.ExperimentItem.ExperimentItemClickedEvent;
-import life.qbic.datamanager.views.support.experiment.ExperimentItemCollection;
-import life.qbic.datamanager.views.support.experiment.ExperimentItemCollection.AddExperimentClickEvent;
-import life.qbic.datamanager.views.support.experiment.ExperimentItemCollection.ExperimentSelectionEvent;
 import life.qbic.projectmanagement.application.ExperimentInformationService;
 import life.qbic.projectmanagement.application.ExperimentalDesignSearchService;
 import life.qbic.projectmanagement.domain.project.ProjectId;
 import life.qbic.projectmanagement.domain.project.experiment.Experiment;
+import life.qbic.projectmanagement.domain.project.experiment.ExperimentId;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
  * Experiment list component
  * <p>
  * The list component is a {@link PageArea} component, which is responsible for showing the
- * {@link Experiment} information for all experiments in its {@link ExperimentItemCollection }
- * within the currently examined {@link life.qbic.projectmanagement.domain.project.Project}.
+ * {@link Experiment} information for all experiments in its {@link ListBox} within the currently
+ * examined {@link life.qbic.projectmanagement.domain.project.Project}.
  * <p>
  * Additionally, it provides the possibility to create new experiments with its
  * {@link ExperimentInformationDialog} and enables the user to select an experiment of interest via
- * clicking on the {@link ExperimentItem} associated with the experiment.
+ * clicking on the item within {@link ListBox} associated with the experiment.
  * <p>
  * Finally, it allows components to be informed about a new experiment creation or selection via the
- * {@link
- * life.qbic.datamanager.views.support.experiment.ExperimentItemCollection.ExperimentSelectionEvent}
- * event.
+ * {@link ExperimentSelectionEvent} event.
  */
 @SpringComponent
 @UIScope
 public class ExperimentListComponent extends PageArea {
+
   @Serial
   private static final long serialVersionUID = -2196400941684042549L;
-  private final ExperimentItemCollection experimentItemCollection;
   private final transient ExperimentInformationService experimentInformationService;
+  private final Disclaimer noExperimentDisclaimer;
+  private final Div header = new Div();
+  private final Div content = new Div();
+  private final ListBox<Experiment> listBox = new ListBox<>();
   private Context context;
-
 
   public ExperimentListComponent(
       @Autowired ExperimentInformationService experimentInformationService,
@@ -53,10 +58,11 @@ public class ExperimentListComponent extends PageArea {
     Objects.requireNonNull(experimentInformationService);
     Objects.requireNonNull(experimentalDesignSearchService);
     this.experimentInformationService = experimentInformationService;
-    this.experimentItemCollection = ExperimentItemCollection.create();
-    this.add(experimentItemCollection);
+    this.add(listBox);
     this.addClassName("experiment-list-component");
-    addItemCollectionListeners();
+    listBox.addClassName("transparent-icons");
+    layoutComponents();
+    this.noExperimentDisclaimer = createNoExperimentDisclaimer();
   }
 
   public void setContext(Context context) {
@@ -64,45 +70,125 @@ public class ExperimentListComponent extends PageArea {
     refresh();
   }
 
-  private void loadExperimentsForProject(ProjectId projectId) {
-    experimentItemCollection.removeAll();
-    Collection<Experiment> foundExperiments = experimentInformationService.findAllForProject(
-        projectId);
-    if (foundExperiments.isEmpty()) {
-      experimentItemCollection.showNoExperimentDisclaimer(true);
-    } else {
-      experimentItemCollection.showNoExperimentDisclaimer(false);
-      foundExperiments.forEach(this::addExperimentToExperimentItemCollection);
-    }
+  private void layoutComponents() {
+    initHeader();
+    content.addClassName("content");
+    add(content);
   }
 
+  private void loadExperimentsForProject(ProjectId projectId) {
+    Collection<Experiment> foundExperiments = experimentInformationService.findAllForProject(
+        projectId);
+    remove(header);
+    remove(content);
+    content.removeAll();
+    if (foundExperiments.isEmpty()) {
+      content.add(noExperimentDisclaimer);
+    } else {
+      addComponentAsFirst(header);
+      setExperimentsInListBox(foundExperiments);
+      content.add(listBox);
+    }
+    add(content);
+  }
 
+  private void setExperimentsInListBox(Collection<Experiment> experiments) {
+    listBox.setItems(experiments);
+    listBox.setRenderer(new ComponentRenderer<Component, Experiment>(
+        experiment -> new Span(experiment.getName())));
+    listBox.addValueChangeListener(
+        listBoxExperimentComponentValueChangeEvent -> {
+          if (listBoxExperimentComponentValueChangeEvent.isFromClient()) {
+            fireExperimentSelectionEvent(
+                listBoxExperimentComponentValueChangeEvent.getValue().experimentId(),
+                listBoxExperimentComponentValueChangeEvent.isFromClient());
+          }
+        });
+  }
 
   public void refresh() {
     loadExperimentsForProject(context.projectId()
         .orElseThrow(() -> new ApplicationException("no project id in context " + context)));
   }
 
-  private void addItemCollectionListeners() {
-    experimentItemCollection.addExperimentSelectionListener(this::fireEvent);
-    experimentItemCollection.addAddButtonListener(this::fireEvent);
+  private void initHeader() {
+    Div controls = new Div();
+    controls.addClassName("controls");
+    Button addButton = new Button("Add");
+    addButton.addClassName("primary");
+    addButton.addClickListener(
+        event -> fireEvent(new AddExperimentClickEvent(this, event.isFromClient())));
+    controls.add(addButton);
+    Span title = new Span("Experiments");
+    title.addClassName("title");
+    header.add(title, controls);
+    header.addClassName("header");
+    addComponentAsFirst(header);
   }
 
-
-  private void addExperimentToExperimentItemCollection(Experiment experiment) {
-    experimentItemCollection.addExperimentItem(ExperimentItem.create(experiment));
+  private Disclaimer createNoExperimentDisclaimer() {
+    var disclaimer = Disclaimer.createWithTitle("Add an experiment",
+        "Get started by adding an experiment", "Add experiment");
+    disclaimer.addDisclaimerConfirmedListener(confirmedEvent -> fireEvent(
+        new AddExperimentClickEvent(this, confirmedEvent.isFromClient())));
+    return disclaimer;
   }
+
 
   public void addAddButtonListener(ComponentEventListener<AddExperimentClickEvent> listener) {
     addListener(AddExperimentClickEvent.class, listener);
   }
 
   /**
-   * Adds the provided component listener for {@link ExperimentItemClickedEvent}
+   * Adds the provided component listener for {@link ExperimentSelectionEvent}
    */
   public void addExperimentSelectionListener(
       ComponentEventListener<ExperimentSelectionEvent> experimentSelectionListener) {
     addListener(ExperimentSelectionEvent.class, experimentSelectionListener);
+  }
+
+  private void fireExperimentSelectionEvent(ExperimentId experimentId, boolean fromClient) {
+    var experimentSelectionEvent = new ExperimentSelectionEvent(this, fromClient, experimentId);
+    fireEvent(experimentSelectionEvent);
+  }
+
+  public static class AddExperimentClickEvent extends ComponentEvent<ExperimentListComponent> {
+
+    /**
+     * Creates a new event using the given source and indicator whether the event originated from
+     * the client side or the server side.
+     *
+     * @param source     the source component
+     * @param fromClient <code>true</code> if the event originated from the client
+     *                   side, <code>false</code> otherwise
+     */
+    public AddExperimentClickEvent(ExperimentListComponent source, boolean fromClient) {
+      super(source, fromClient);
+    }
+  }
+
+  public static class ExperimentSelectionEvent extends ComponentEvent<ExperimentListComponent> {
+
+    private final ExperimentId experimentId;
+
+    /**
+     * Creates a new event using the given source and indicator whether the event originated from
+     * the client side or the server side.
+     *
+     * @param source       the source component
+     * @param fromClient   <code>true</code> if the event originated from the client
+     *                     side, <code>false</code> otherwise
+     * @param experimentId the selected experiment
+     */
+    public ExperimentSelectionEvent(ExperimentListComponent source, boolean fromClient,
+        ExperimentId experimentId) {
+      super(source, fromClient);
+      this.experimentId = experimentId;
+    }
+
+    public ExperimentId getExperimentId() {
+      return experimentId;
+    }
   }
 
 }
