+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/projects/project/info/ProjectSupportComponent.java
@@ -8,8 +8,8 @@ import java.io.Serial;
 import java.util.Objects;
 import life.qbic.datamanager.views.Context;
 import life.qbic.datamanager.views.projects.project.experiments.ExperimentListComponent;
-import life.qbic.datamanager.views.support.experiment.ExperimentItemCollection.AddExperimentClickEvent;
-import life.qbic.datamanager.views.support.experiment.ExperimentItemCollection.ExperimentSelectionEvent;
+import life.qbic.datamanager.views.projects.project.experiments.ExperimentListComponent.AddExperimentClickEvent;
+import life.qbic.datamanager.views.projects.project.experiments.ExperimentListComponent.ExperimentSelectionEvent;
 import life.qbic.logging.api.Logger;
 import life.qbic.logging.service.LoggerFactory;
 import life.qbic.projectmanagement.domain.project.experiment.Experiment;
